requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.6
django-heroku==0.3.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
gunicorn==20.1.0
psycopg2==2.9.1
PyJWT==2.1.0
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.3.0